+++ examples/example_test.go
@@ -27,31 +27,10 @@ func TestIntegration(t *testing.T) {
 			t.Log("OK!")
 		})
 	}
-	//_ = service.TopicService.Remove("dispatcher.example.testing", "dispatcher.example.testing_ERR", "dispatcher.example.testing_Reply")
 }
 
 func BenchmarkProducer(b *testing.B) {
 	for i := 0; i < b.N; i++ {
 		send(topic, 1)
-	}
-}
-
-func TestMultiConsProds(t *testing.T) {
-	type args struct {
-		msgCount int
-	}
-	tests := []struct {
-		name string
-		args args
-		want int
-	}{
-		{"10 Messages", args{10}, 20},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			if got := MultiPubSubs(tt.args.msgCount); got != tt.want {
-				t.Errorf("MultiPubSubs() = %v, want %v", got, tt.want)
-			}
-		})
 	}
 }